+++ frontend/packages/react-app/package.json
@@ -25,6 +25,7 @@
     "graphql": "^14.6.0",
     "react": "16.12.0",
     "react-dom": "16.12.0",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "^3.3.1",
     "styled-components": "^5.1.0"
   },
